+++ src/Client/TimeSpan.ts
@@ -1,0 +1,55 @@
+/**
+ * Provides some common time intervals in milliseconds.
+ */
+export const TimeSpan = Object.freeze({
+
+	/**
+	 * The number of hours in 1 day.
+	 */
+	HoursPerDay: 24,
+
+	/**
+	 * The number of minutes in 1 day.
+	 */
+	MinutesPerDay: 1_440,
+
+	/**
+	 * The number of minutes in 1 hour.
+	 */
+	MinutesPerHour: 60,
+
+	/**
+	 * The number of seconds in 1 day.
+	 */
+	SecondsPerDay: 86_400,
+
+	/**
+	 * The number of seconds in 1 hour.
+	 */
+	SecondsPerHour: 3_600,
+
+	/**
+	 * The number of seconds in 1 minute.
+	 */
+	SecondsPerMinute: 60,
+
+	/**
+	 * The number of milliseconds in 1 day.
+	 */
+	MillisecondsPerDay: 86_400_000,
+
+	/**
+	 * The number of milliseconds in 1 hour.
+	 */
+	MillisecondsPerHour: 3_600_000,
+
+	/**
+	 * The number of milliseconds in 1 minute.
+	 */
+	MillisecondsPerMinute: 60_000,
+
+	/**
+	 * The number of milliseconds in 1 second.
+	 */
+	MillisecondsPerSecond: 1_000
+});