+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Version [0.27.0](https://github.com/cedx/base/compare/v0.20.0...v0.27.0)
 - Added the `today()` function.
+- Replaced the `Duration` enumeration by the `TimeSpan` object.
 
 ## Version [0.26.0](https://github.com/cedx/base/compare/v0.25.0...v0.26.0)
 - Added the `Data.Mapping.MailAddressTypeHandler` class.